Squ1rrel_CTF_2026/squ1rrel2026CTF/yolo/starting_files/extracted/requirements.txt
flask==3.0.3
gunicorn==22.0.0
numpy==1.23.5
opencv-python-headless==4.9.0.80
Pillow==10.3.0
PyYAML==6.0.1
matplotlib==3.7.5
scipy==1.10.1
tqdm==4.66.4
pandas==2.0.3
seaborn==0.13.2
requests==2.32.3
protobuf==3.20.3
psutil==5.9.8
torch==1.13.1
torchvision==0.14.1
